Add files via upload
The basic mod that adds 5 new Xenos:
- a black basic xeno,
- a melee centric xeno,
- a beautiful leader, 
- an animal-based xeno,
- a less immediately dangerous Sanguophage,

as well as 4 factions:
two tribal-level factions and 2 outlander-level factions. One regular and the other a natural enemy. 

Future modules will add more Xenotypes, factions and individual gimmicks. 

Currently suggested mods to play with are Roo's Minotaur and Satyr mods for the Wildlox, VRE Highmate for Pawloc and Headlox, and Sanguophage for the Hemoloc.